requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
pydantic==2.4.2
starlette==0.27.0
requests==2.31.0
strawberry-graphql==0.211.1
owlready2==0.44
pyinstaller==6.4.0